fixing bugs record data and index

+++ src/main/java/com/aritmetic/op/api/dtos/RecordResponseDto.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Builder
 public class RecordResponseDto {
     private List<Record> listOfRecords;
-    private int totalRecords;
+    private long totalRecords;
 }
